backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["HeskethGD <graham.hesketh@opsmorph.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
streamlit = "^1.40.1"
fastapi = "^0.115.5"
lancedb = "^0.16.0"
openai = "^1.55.0"
pymupdf = "^1.24.14"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
python-dotenv = "^1.0.1"
ollama = "^0.3.3"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
